feat: 장기판 기물 초기화 기능 구현
