+++ README.md
@@ -26,11 +26,12 @@
     - [x] 진영을 가진다.
     - [x] 위치를 가진다.
     - 움직일 수 있다.
-      - [x] [예외] 기물별 규칙을 벗어난 움직임을 제한한다.
-      - [x] [예외] 기물의 현재위치로 이동을 제한한다.
-      - [x] [예외] 이동 위치 경로상 다른 기물이 존재할 경우 이동을 제한한다.
+        - [x] [예외] 기물별 규칙을 벗어난 움직임을 제한한다.
+        - [x] [예외] 기물의 현재위치로 이동을 제한한다.
+        - [x] [예외] 이동 위치 경로상 다른 기물이 존재할 경우 이동을 제한한다.
 
 2. 위치
+    - [x] 10 * 9 (세로, 가로) 영역을 가진다.
     - [x] 가로, 세로 2차원 좌표를 가진다.
     - [x] 좌표를 변경할 수 있다.
     - [x] [예외] 좌표의 범위는 세로 10, 가로 9로 제한한다.
@@ -39,7 +40,6 @@
     - [x] 특정 x, y 좌표와 같은지 비교한다.
 
 3. 장기판
-    - 10 * 9 (세로, 가로) 영역을 가진다.
-    - 각 진영 별로 기물들의 위치를 초기화시킨다.
+    - [x] 각 진영 별로 기물들의 위치를 초기화시킨다.
     - 각 진영 별로 기물을 움직일 수 있다.
         - 움직인 위치에 다른 기물이 있으면 해당 기물을 제거한다.
